% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apifunc.R
\name{geo_api}
\alias{geo_api}
\title{Geocoding response from Google Map Api}
\usage{
geo_api(address)
}
\arguments{
\item{...}{adress of any place aound the globe.}

\item{valid}{adress name of a place.}
}
\value{
All the contents of the adress speciefied as input, 
       including the Geocoding longitude and latitude of the given place.
  \url{https://developers.google.com/maps/}
}
\description{
\code{geo_api function} returns the geocoding responses of it's argument(given adress).
}
\details{
This is a Web Api funntion which gets geographical information of a given adress from the Google Map.
}
\examples{
geo_api("Newyork")
geo_api("Asmara")

}

