% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\name{GetAssetsInfo}
\alias{GetAssetsInfo}
\title{Get Asset Info}
\arguments{
\item{id}{a bool. Assets id (optional)}
}
\value{
data.table with asset info

data.table with asset info
}
\description{
Get Asset Info

Get Position Info
}
