% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWebClient.R
\docType{class}
\name{RTTWebApiHost}
\alias{RTTWebApiHost}
\title{RTTWebApiHost}
\description{
RTTWebApiHost
}
\section{Fields}{

\describe{
\item{\code{client.}}{RTTWebClient obj.}
}}

\section{Methods}{

\describe{
\item{\code{GetAccountDetails(id)}}{Get Position Info}

\item{\code{GetAccountInfo()}}{Get Position Info}

\item{\code{GetAccountLogin(id)}}{Get Position Info}

\item{\code{GetAccountWebToken(id)}}{Get Position Info}

\item{\code{GetAllAccountsInfo()}}{Get Position Info}

\item{\code{GetAssetsInfo(id = NULL)}}{Get Asset Info}

\item{\code{GetBarsHistory(
  symbol,
  barsType = "Bid",
  periodicity = "M1",
  startTime,
  endTime = as.POSIXct(Sys.Date(), tz = "GMT"),
  count = 0L
)}}{Get Bar History}

\item{\code{GetCurrencyInfo()}}{Get All Currencies}

\item{\code{GetCurrentQuotes()}}{Get All Current Quotes}

\item{\code{GetDividends()}}{Get All Dividends}

\item{\code{GetGrossTradesByAccountId(id)}}{Get Position Info}

\item{\code{GetGrossTradesByTradeId(id)}}{Get Position Info}

\item{\code{GetManagerAccountSnapshots(
  loginVector,
  from,
  to,
  pagingDirection = as.character(NA),
  pagingSize = 100,
  groupFilter = character(0)
)}}{Get Account Snapshots Info}

\item{\code{GetManagerTradeHistory(
  loginVector,
  from,
  to,
  pagingDirection = as.character(NA),
  pagingSize = 100,
  orderId = as.character(NA),
  skipCancelOrder = TRUE
)}}{Get TradeHistoryReport Info}

\item{\code{GetNetTradesByAccountId(id)}}{Get Position Info}

\item{\code{GetNetTradesByTradeId(id)}}{Get Position Info}

\item{\code{GetPipsValue(targetCurrency, symbols)}}{Get Pips Value}

\item{\code{GetPositionInfo(id = NULL)}}{Get Position Info}

\item{\code{GetSplits(symbolFilter = NULL)}}{Get Split Info}

\item{\code{GetSymbolsInfo(getPipsValue = FALSE, getLastPrice = TRUE)}}{Get All Symbols}

\item{\code{GetTickHistory(
  symbol,
  startTime,
  endTime = as.POSIXct(Sys.Date(), tz = "GMT"),
  count = 0L
)}}{Get Bar History}

\item{\code{GetTradeInfo(id = NULL)}}{Get Trade Info}
}}

