% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagedoc.R
\docType{package}
\name{RWebApi}
\alias{RWebApi}
\alias{RWebApi-package}
\title{Geo Web Api: A package that takes address and gives out Longitude, Latitude and the Name of the location, by incorporating with the google geo codes .}
\arguments{
\item{...}{adress of any place aound the globe.}

\item{valid}{adress name of a place.}
}
\value{
All the contents of the adress speciefied as input, 
       including the Geocoding longitude and latitude of the given place.
  \url{https://developers.google.com/maps/}
}
\description{
The Geo Web Api package provides one function called geo_api.
}
\section{geo_api function}{
it takes address as a string, and retruns coordinates and location name as per google geo codes.
}
\examples{
geo_api("Newyork")
geo_api("Asmara")
}

